installed tailwind, refactor navbar
