+++ src/components/landing/index.js
@@ -3,5 +3,3 @@ export { default as ChartVisual } from "./ChartVisual";
 export { default as MapVisual } from "./MapVisual";
 export { default as TableVisual } from "./TableVisual";
 export { default as UserVisual } from "./UserVisual";
-
-export { default as LandingPage } from "./LandingPage";
