+++ src/components/pages/AddStorePage.js
@@ -1,0 +1,11 @@
+import React from "react"
+
+const AddStorePage = () => {
+ return (
+   <div>
+     <h2>Form Header</h2>
+   </div>
+ )
+}
+
+export default AddStorePage
