+++ src/components/pages/index.js
@@ -1,0 +1,2 @@
+export { default as LandingPage } from "./LandingPage";
+export { default as AddStorePage } from "./AddStorePage";
