+++ src/styles/global/App.css
@@ -1,1 +1,5 @@
-
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
