--- src/styles/global/index.css
@@ -1,5 +1,0 @@
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
