+++ src/styles/landing/MapVisual.css
@@ -30,7 +30,10 @@
 }
 
 .fruit-marker-container {
-  margin: 3rem 0 0.5rem 0;
+  display:flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  margin: 3rem 0 1rem;
 }
 
 .fruit-marker-container img {
